address CodeRabbit review on #215
- bot_test.go: TestBotHasNoCrossRoundState now reuses an already-driven Bot
  for its first comparison arm instead of two never-yet-used instances —
  two fresh instances driven through the same view sequence would evolve
  identically even with hidden cross-round state, since neither has any
  prior history to diverge from; only reusing an instance across two drives
  exposes that.
- bot.go: Tier's per-constant doc comments now state RFC-001 §14.3's target
  behaviour as planned, not current — every tier's Decide is presently the
  same provisional decideStayOrMove, and the comments previously read as
  describing what's already implemented.
- rng.go: NewBotRNG encodes seat as a full uint64 rather than truncating to
  uint32. game.SeatID is a plain int with no enforced upper bound in
  Config.Validate, so the narrower encoding could in principle collide two
  distinct seats onto the same bot stream — exactly what this derivation
  exists to prevent (RFC-001 §14.5).

+++ internal/rules/rng.go
@@ -103,8 +103,14 @@ type BotRNG struct{ r *RNG }
 // per-round independence for a fixed seat falls out of reusing it unchanged;
 // the only new derivation here is per-seat, which two bot seats must never
 // share (RFC-001 §14.5, D32).
+//
+// seat is encoded as a full uint64, not truncated to 32 bits: game.SeatID is
+// a plain int (platform width), and Config.Validate places no upper bound on
+// it for a custom player count, so a narrower encoding could in principle
+// collide two distinct seats onto the same stream — exactly the non-
+// collusion property this derivation exists to guarantee.
 func NewBotRNG(matchSeed [32]byte, seat game.SeatID, round int) *BotRNG {
-	msg := binary.BigEndian.AppendUint32([]byte("cinzal.bot.rng"), uint32(seat))
+	msg := binary.BigEndian.AppendUint64([]byte("cinzal.bot.rng"), uint64(seat))
 
 	mac := hmac.New(sha256.New, matchSeed[:])
 	mac.Write(msg)
